Move factory to module
